+++ BackEnd/MyBudgetIA/MyBudgetIA.Api/Controllers/BaseController.cs
@@ -14,10 +14,16 @@ namespace MyBudgetIA.Api.Controllers
         protected IActionResult OkResponse(string message = "Success")
             => Ok(ApiResponse.Ok(message));
 
-        protected IActionResult FailResponse<T>(string message, ApiError[]? errors = null)
+        protected IActionResult BadRequestResponse<T>(string message, ApiError[]? errors = null)
             => BadRequest(ApiResponse<T>.Fail(message, errors));
 
-        protected IActionResult FailResponse(string message, ApiError[]? errors = null)
+        protected IActionResult BadRequestResponse(string message, ApiError[]? errors = null)
             => BadRequest(ApiResponse.Fail(message, errors));
+
+        protected IActionResult NotFoundResponse<T>(string message, ApiError[]? errors = null)
+            => NotFound(ApiResponse<T>.Fail(message, errors));
+
+        protected IActionResult NotFoundResponse(string message, ApiError[]? errors = null)
+            => NotFound(ApiResponse.Fail(message, errors));
     }
 }